# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AskContext": {
        "additionalProperties": false,
        "properties": {
          "active_panel": {
            "anyOf": [
              {
                "enum": [
                  "twins",
                  "shift",
                  "explain",
                  "recipe",
                  "ask"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Panel"
          },
          "selected_dish_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Dish Id"
          }
        },
        "title": "AskContext",
        "type": "object"
      },
      "AskRequest": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/AskContext"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "messages"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskResponse": {
        "additionalProperties": false,
        "properties": {
          "grounded": {
            "description": "false = reply replaced by a template because it cited numbers the engine did not return",
            "title": "Grounded",
            "type": "boolean"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "tool_trace": {
            "items": {
              "$ref": "#/components/schemas/ToolCallTrace"
            },
            "title": "Tool Trace",
            "type": "array"
          },
          "ui_actions": {
            "items": {
              "discriminator": {
                "mapping": {
                  "explain_pair": "#/components/schemas/ExplainPair",
                  "highlight_dishes": "#/components/schemas/HighlightDishes",
                  "open_panel": "#/components/schemas/OpenPanel",
                  "select_dish": "#/components/schemas/SelectDish",
                  "set_shift": "#/components/schemas/SetShift",
                  "show_twins": "#/components/schemas/ShowTwins"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SelectDish"
                },
                {
                  "$ref": "#/components/schemas/ShowTwins"
                },
                {
                  "$ref": "#/components/schemas/SetShift"
                },
                {
                  "$ref": "#/components/schemas/ExplainPair"
                },
                {
                  "$ref": "#/components/schemas/OpenPanel"
                },
                {
                  "$ref": "#/components/schemas/HighlightDishes"
                }
              ]
            },
            "title": "Ui Actions",
            "type": "array"
          }
        },
        "required": [
          "reply",
          "grounded",
          "ui_actions",
          "tool_trace"
        ],
        "title": "AskResponse",
        "type": "object"
      },
      "ChatMessage": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExplainPair": {
        "additionalProperties": false,
        "properties": {
          "a": {
            "title": "A",
            "type": "string"
          },
          "b": {
            "title": "B",
            "type": "string"
          },
          "type": {
            "const": "explain_pair",
            "default": "explain_pair",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "a",
          "b"
        ],
        "title": "ExplainPair",
        "type": "object"
      },
      "ExplainResponse": {
        "additionalProperties": false,
        "properties": {
          "a": {
            "title": "A",
            "type": "string"
          },
          "attributions": {
            "$ref": "#/components/schemas/PairAttributions"
          },
          "b": {
            "title": "B",
            "type": "string"
          },
          "dims": {
            "items": {
              "$ref": "#/components/schemas/DimComparison"
            },
            "title": "Dims",
            "type": "array"
          },
          "distance": {
            "title": "Distance",
            "type": "number"
          },
          "similarity_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Similarity Pct",
            "type": "number"
          }
        },
        "required": [
          "a",
          "b",
          "similarity_pct",
          "distance",
          "dims",
          "attributions"
        ],
        "title": "ExplainResponse",
        "type": "object"
      },
...
      "HighlightDishes": {
        "additionalProperties": false,
        "properties": {
          "dish_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Dish Ids",
            "type": "array"
          },
          "type": {
            "const": "highlight_dishes",
            "default": "highlight_dishes",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "dish_ids"
        ],
        "title": "HighlightDishes",
        "type": "object"
      },
...
      "RecipeRequest": {
        "additionalProperties": false,
        "properties": {
          "course": {
            "anyOf": [
              {
                "enum": [
                  "savory",
                  "dessert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course"
          },
          "format": {
            "anyOf": [
              {
                "enum": [
                  "soup",
                  "noodles",
                  "rice_dish",
                  "stew_braise",
                  "fried",
                  "grilled_roasted",
                  "raw_salad",
                  "bread_sandwich",
                  "dumpling",
                  "baked_sweet",
                  "custard_creamy",
                  "frozen"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "RecipeRequest",
        "type": "object"
      },
      "RecipeResponse": {
        "additionalProperties": false,
        "properties": {
          "attributions": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/Contribution"
              },
              "type": "array"
            },
            "propertyNames": {
              "enum": [
                "sweet",
                "salty",
                "sour",
                "bitter",
                "umami",
                "spicy",
                "smoky",
                "roasted",
                "fermented",
                "herbal",
                "fruity",
                "earthy",
                "rich",
                "creamy",
                "crispy",
                "chewy",
                "brothy"
              ]
            },
            "title": "Attributions",
            "type": "object"
          },
          "course": {
            "enum": [
              "savory",
              "dessert"
            ],
            "title": "Course",
            "type": "string"
          },
          "coverage": {
            "$ref": "#/components/schemas/Coverage"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/RecipeLine"
            },
            "title": "Lines",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "neighbors": {
            "items": {
              "$ref": "#/components/schemas/NeighborResult"
            },
            "title": "Neighbors",
            "type": "array"
          },
          "used_grok": {
            "title": "Used Grok",
            "type": "boolean"
          },
          "vector": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "sweet",
                "salty",
                "sour",
                "bitter",
                "umami",
                "spicy",
                "smoky",
                "roasted",
                "fermented",
                "herbal",
                "fruity",
                "earthy",
                "rich",
                "creamy",
                "crispy",
                "chewy",
                "brothy"
              ]
            },
            "title": "Vector",
            "type": "object"
          },
          "xyz": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Xyz",
            "type": "array"
          }
        },
        "required": [
          "name",
          "course",
          "lines",
          "coverage",
          "used_grok",
          "vector",
          "xyz",
          "neighbors",
          "attributions"
        ],
        "title": "RecipeResponse",
        "type": "object"
      },
...
      "ShiftRequest": {
        "additionalProperties": false,
        "properties": {
          "deltas": {
            "additionalProperties": {
              "description": "sigma units",
              "maximum": 2.0,
              "minimum": -2.0,
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "sweet",
                "salty",
                "sour",
                "bitter",
                "umami",
                "spicy",
                "smoky",
                "roasted",
                "fermented",
                "herbal",
                "fruity",
                "earthy",
                "rich",
                "creamy",
                "crispy",
                "chewy",
                "brothy"
              ]
            },
            "title": "Deltas",
            "type": "object"
          },
          "dish_id": {
            "title": "Dish Id",
            "type": "string"
          },
          "k": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "K",
            "type": "integer"
          }
        },
        "required": [
          "dish_id"
        ],
        "title": "ShiftRequest",
        "type": "object"
      },
      "ShiftResponse": {
        "additionalProperties": false,
        "properties": {
          "relaxation_level": {
            "minimum": 0.0,
            "title": "Relaxation Level",
            "type": "integer"
          },
          "relaxation_note": {
            "title": "Relaxation Note",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ShiftResult"
            },
            "title": "Results",
            "type": "array"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "target_xyz": {
            "maxItems": 3,
            "minItems": 3,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Target Xyz",
            "type": "array"
          },
          "target_z": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "sweet",
                "salty",
                "sour",
                "bitter",
                "umami",
                "spicy",
                "smoky",
                "roasted",
                "fermented",
                "herbal",
                "fruity",
                "earthy",
                "rich",
                "creamy",
                "crispy",
                "chewy",
                "brothy"
              ]
            },
            "title": "Target Z",
            "type": "object"
          }
        },
        "required": [
          "source_id",
          "target_z",
          "target_xyz",
          "relaxation_level",
          "relaxation_note",
          "results"
        ],
        "title": "ShiftResponse",
        "type": "object"
      },
...
      "SpaceResponse": {
        "additionalProperties": false,
        "properties": {
          "cuisines": {
            "items": {
              "$ref": "#/components/schemas/CuisineMeta"
            },
            "title": "Cuisines",
            "type": "array"
          },
          "dims": {
            "items": {
              "$ref": "#/components/schemas/DimMeta"
            },
            "title": "Dims",
            "type": "array"
          },
          "dishes": {
            "items": {
              "$ref": "#/components/schemas/DishPoint"
            },
            "title": "Dishes",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/SpaceMeta"
          },
          "pca": {
            "$ref": "#/components/schemas/PcaInfo"
          }
        },
        "required": [
          "dims",
          "cuisines",
          "dishes",
          "pca",
          "meta"
        ],
        "title": "SpaceResponse",
        "type": "object"
      },
...
      "TwinsResponse": {
        "additionalProperties": false,
        "properties": {
          "relaxation_level": {
            "description": "0 = strictest threshold met",
            "minimum": 0.0,
            "title": "Relaxation Level",
            "type": "integer"
          },
          "relaxation_note": {
            "title": "Relaxation Note",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "twins": {
            "items": {
              "$ref": "#/components/schemas/TwinResult"
            },
            "title": "Twins",
            "type": "array"
          }
        },
        "required": [
          "source_id",
          "relaxation_level",
          "relaxation_note",
          "twins"
        ],
        "title": "TwinsResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "TasteSpace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ask": {
      "post": {
        "operationId": "postAsk",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Post Ask"
      }
    },
    "/api/explain": {
      "get": {
        "operationId": "getExplain",
        "parameters": [
          {
            "in": "query",
            "name": "a",
            "required": true,
            "schema": {
              "title": "A",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "b",
            "required": true,
            "schema": {
              "title": "B",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Explain"
      }
    },
...
    "/api/recipe": {
      "post": {
        "operationId": "postRecipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Post Recipe"
      }
    },
    "/api/shift": {
      "post": {
        "operationId": "postShift",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Post Shift"
      }
    },
    "/api/space": {
      "get": {
        "operationId": "getSpace",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Space"
      }
    },
    "/api/twins": {
      "get": {
        "operationId": "getTwins",
        "parameters": [
          {
            "description": "dish id",
            "in": "query",
            "name": "dish",
            "required": true,
            "schema": {
              "description": "dish id",
              "title": "Dish",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "k",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 10,
              "minimum": 1,
              "title": "K",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwinsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Twins"
      }
    }
  }
}